pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bili-pipeline"
version = "0.1.0"
description = "Minimal runnable scaffold for Bilibili video pool discovery."
readme = "methods-v1.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp",
    "bilibili-api-python",
    "openpyxl>=3.1.5",
    "oss2",
    "pandas",
    "streamlit",
    "xlrd>=2.0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
